Introduce EnumParams, parsed structs
